Uploaded files
Working on trying to add multiple drag and drop elements and have the items being dragged remain in their current position. Essentially, I want their class name to be copied on the empty containers upon dragDrop. However, at this time I am only able to get the quarter note to drag to the location (it is not copied, it just leaves its home base.